Move and hide checkTruncated() method to leave ArrayDataInput interface unperturbed.
