+++ src/test/java/nom/tam/fits/TruncatedFileExceptionTest.java
@@ -32,8 +32,12 @@ package nom.tam.fits;
  */
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
+
+import nom.tam.util.FitsFile;
 
 public class TruncatedFileExceptionTest {
 
@@ -41,5 +45,17 @@ public class TruncatedFileExceptionTest {
     public void testConstructorWithCause() {
         Exception e = new TruncatedFileException("test exception", new IllegalArgumentException("test cause"));
         assertEquals(IllegalArgumentException.class, e.getCause().getClass());
+    }
+    
+    @Test
+    public void testTruncated() throws Exception {
+        try (FitsFile f = new FitsFile("fftest.bin", "rw", 100)) {
+            f.seek(10);
+            f.setLength(5);
+            assertEquals(5, f.getFilePointer());
+            assertFalse(Fits.checkTruncated(f));
+            f.seek(10);
+            assertTrue(Fits.checkTruncated(f));
+        }
     }
 }
